requirements.txt
requests==2.31.0
beautifulsoup4==4.13.4
sentence-transformers==2.2.2
chromadb==0.4.24
flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2